refactor: split MarkdownReporter into a package
